refactor(library/math): moves "greatestCommonDivisor.ts" to "greatestCommonDivisor/abbreviations/gcf.ts"
