+++ src/library/math/operator/constructive/exports.ts
@@ -1,6 +1,6 @@
 export * from './absoluteValueOf/aliases/modulusOf';
 
-export * from './greatestCommonDivisor';
+export * from './greatestCommonDivisor/abbreviations/gcf';
 
 export * from './leastCommonMultiple';
